Admin Panel design updated
Having an issue with the Dropdown list. after a join of SQL Table at the "Product Controller" is sending only one content in a list. [line 33] . in there we are using a custom datatype which is at Model/CreateFormViewModel.cs . If any further error occur this text might be useful.